requirements.txt
kafka-python
pyspark
pandas
